refactor: InputValidator를 model 디렉토리 하위로 옮기고 LottoValidator라 변경하다
